use TerminalViewManager to manage terminal views
Right now, some properties are store is a view object which is not managed by
TerminalView but [sublime_plugin][1]. This is not satifactory, so I added a
TerminalViewManager to store the TerminalView instance.

The instance can be retrieved later for later purposes. Ultimately, the
`view.foo` should be replace by `TerminalViewManager.load_from_id(vid).foo`.

[1]: https://github.com/twolfson/sublime-files/blob/afeed6b8f75ca07ed7f1de80d10a9a2a12764e06/sublime_plugin.py#L196