Catch FileNotFoundError when opening TerminalView from e.g. a temporary file (#47)
* Catch FileNotFoundError in case of a faulty cwd

The directory of the file_name() of the current view may not exist, so we have
to account for that. For example, opening a file from a zipfile (e.g. via
the PackageResourceViewer plugin), and then starting a terminal results in
this exception being thrown.

* Switch around the construction of SublimeTerminalBuffer and LinuxPty

This is so that we don't have to deregister the SublimeTerminalBuffer from
its manager in case of a FileNotFoundError.

* Revert to home dir instead of showing error