Allow cmd argument to be string (split by shlex) or list of strings (#63)
* Use shlex to split the command arguments

This allows proper use of quotes in the command. For example, using

/bin/bash -c 'ssh example.com'

is now split into

- "/bin/bash"
- "-c"
- "ssh example.com"

which allows for extra arguments after the `-c` option of bash.

* Make a "list of strings" for the cmd the default behavior

This is list of strings is passed to LinuxPty and handled from there.
If "cmd" is a string, it is split with shlex.split to account for single
and double quotes. These changes are also reflected in the exec command,
where a `"shell_cmd"` argument is split with shlex, and a `"cmd"`
argument is used as-is.

* Add missing import in exec.py

We need this import for showing an error message to the user.

* Remove old catch guard

This was present because I attempted to check if cmd[0] actually exists and
if not, raise a KeyError. It turned out that this was harder to implement
than I thought.